+++ HISTORY.md
@@ -1,5 +1,21 @@
 # Revision History for `matdb`
 
+## Revision 0.0.6
+
+- Replaced self.collections with self.seeded in `matdb/database/controller.py`.
+- Assigned self.steps to the Sequence database.
+- Added `support/matdb_sbatch.py` for the unit testing stub for sbatch.
+- Added `matdb/querries.py` so that the different environments, i.e.,
+  sbatch and Tracy, can be handled appropriately.
+- Set the `self._bands` attribute in the phonon database.
+- Added `EnumDatabase` for generating enumerated runs.
+- Added `lattice.in` template for phenum's use.
+- Fixed the `fileformat` in `matdb/kpoints.py` to be `vasp-ase` to
+  match materialscloud server update.
+- Changed `matdb/plotting/potentials.py` so that if the
+  `atoms.config_type` attribute does not exist it doesn't throw
+  errors.
+
 ## Revision 0.0.5
 
 - Added `DynamicDatabase` for generating MD runs.